app/pyproject.toml
[project]
name = "employee-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "openai>=1.51",          # SDK v1
    "streamlit>=1.37",
    "gradio>=4.44",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.9",
    "httpx>=0.27",
]

[tool.uv]
index-url = "https://pypi.org/simple"
